Add ".." as a dedicated concat operator
Instead of overloading "+", add ".." as a dedicated operator for concat.
I prefer Lua's two dots over PHP's one dot and this is also easier to
implement I think. Right now only numbers and strings are supported, but
you can concat two numbers if you want and get a string output. This is
a more consistent language design than being able to concat one string
and one number and have it return a string as suggested in one of the
challenges I think, and I like having a dedicated concat operator.