Add support for expressions evaluating in REPL
This is exercise 8.1.

The list of statements is now a class property so that we can check
its length to ensure that there is actually only a single expression
in our list when we evaluate.

I'm not sure if this is the right approach; from the grammar I think
this should be evaluated very first, but I couldn't figure out how to
do it from that approach. (I spent a REALLY long time figuring
this out.)