Add support for break statements
This is exercise 9.3. Technically at that time, the resolver
didn't exist, but I did this after reading chapter 11, so I figured
I may as well put the error handling in the right place. Also I did
figure out how to do this (i.e. using an exception to handle the
change in control flow) prior to reading about returns.